Sidebar component BEMised
