+++ styles/pages/_chapter.scss
@@ -1,3 +1,4 @@
 @use '../tools' as *;
 @use '../../components/navigation/navigatiom';
 @use '../../components/menu/menu';
+@use '../../components/sidebar/sidebar';
